+++ smoltrace/__init__.py
@@ -6,11 +6,18 @@ __version__ = "0.1.1"
 
 # Export main functions
 from .core import run_evaluation
-from .utils import cleanup_datasets, discover_smoltrace_datasets, filter_runs, group_datasets_by_run
+from .utils import (
+    cleanup_datasets,
+    compute_pass_at_1,
+    discover_smoltrace_datasets,
+    filter_runs,
+    group_datasets_by_run,
+)
 
 __all__ = [
     "run_evaluation",
     "cleanup_datasets",
+    "compute_pass_at_1",
     "discover_smoltrace_datasets",
     "group_datasets_by_run",
     "filter_runs",